refactor: replace dimensions and spaces modules
