+++ mod.rs
@@ -1,0 +1,5 @@
+/// 1d array type.
+pub type Vector<T = f64> = ndarray::Array1<T>;
+
+/// 2d array type.
+pub type Matrix<T = f64> = ndarray::Array2<T>;
