+++ rustfmt.toml
@@ -5,6 +5,5 @@ imports_layout = "HorizontalVertical"
 match_block_trailing_comma = true
 normalize_comments = true
 reorder_imports = true
-reorder_imports_in_group = true
 use_try_shorthand = true
 wrap_comments = true
