+++ src/continuous/mod.rs
@@ -1,0 +1,2 @@
+import_all!(continuous);
+import_all!(infinite);
