+++ src/discrete/binary.rs
