+++ src/discrete/mod.rs
@@ -1,0 +1,4 @@
+import_all!(binary);
+import_all!(discrete);
+import_all!(natural);
+import_all!(partitioned);
