+++ src/discrete/natural.rs
