+++ src/product/mod.rs
@@ -1,0 +1,5 @@
+//! Combination space types module.
+
+import_all!(pair);
+import_all!(regular);
+import_all!(named);
