+++ src/product/regular.rs
@@ -1,7 +1,8 @@
-use dimensions::{Continuous, Partitioned};
+use continuous::Continuous;
+use core::{Space, Card, Surjection};
+use discrete::Partitioned;
 use rand::Rng;
 use std::{iter::FromIterator, ops::{Add, Index}, slice::Iter as SliceIter};
-use {Space, Card, Surjection};
 
 /// N-dimensional homogeneous space.
 #[derive(Clone, Serialize, Deserialize, Debug)]
@@ -109,11 +110,16 @@ impl<D: Space> Add<RegularSpace<D>> for RegularSpace<D> {
 
 #[cfg(test)]
 mod tests {
-    use dimensions::{Continuous, Discrete};
-    use ndarray::arr1;
+    extern crate ndarray;
+
+    use continuous::Continuous;
+    use core::{Space, Card, Surjection};
+    use discrete::Discrete;
+    use product::RegularSpace;
     use rand::thread_rng;
+    use self::ndarray::arr1;
+    use std::collections::HashMap;
     use std::iter::FromIterator;
-    use {RegularSpace, Space, Card, Surjection};
 
     #[test]
     fn test_dim() {
